waila_python_backend/requirements.txt
langchain 
langchain_core 
langchain-huggingface 
langchain_experimental 
langchain-community 
langchainhub
langchain-mistralai
python-dotenv
fastapi[standard]
itsdangerous